feat: refactor feedback components
