+++ src/components/ui/answer-feedback.tsx
@@ -1,9 +1,129 @@
-import { CopyToClipboardButton } from "@/components/ui/copy-to-clipboard-button";
+import { texts } from "@/lib/texts";
+import { cn } from "@/lib/utils";
+import React, { useState } from "react";
+import { ThumbsUpIcon } from "./icons/thumbs-up";
+import { ThumbsDownIcon } from "./icons/thumbs-down";
+import { ThumbsDownSolidIcon } from "./icons/thumbs-down-solid";
+import { ThumbsUpSolidIcon } from "./icons/thumbs-up-solid";
+import { Transition } from "@headlessui/react";
+import { XIcon } from "lucide-react";
+import { CopyToClipboardButton } from "./copy-to-clipboard-button";
 
-export function AnswerFeedback({ generatedAnswer }: { generatedAnswer: string }) {
+export function AnswerFeedback({
+	generatedAnswer,
+}: {
+	generatedAnswer: string;
+}) {
+	const [isThumbsUpClicked, setIsThumbsUpClicked] = useState(false);
+	const [isThumbsDownClicked, setIsThumbsDownClicked] = useState(false);
+	const [isThankYouMessageVisible, setIsThankYouMessageVisible] =
+		useState(false);
+	const [areTagsVisible, setAreTagsVisible] = useState(false);
+	const [selectedTag, setSelectedTag] = useState(null);
+	const [isTagDisabled, setIsTagDisabled] = useState(false);
 
-	return <div className="flex justify-between w-full px-4 pt-2">
-		<p>Placeholder</p>
-		<CopyToClipboardButton generatedAnswer={generatedAnswer} />
-	</div>
+	const onThumbsDownClick = () => {
+		setIsThumbsDownClicked(true);
+		setAreTagsVisible(true);
+	};
+
+	const onThumbsUpClick = () => {
+		setIsThumbsUpClicked(true);
+		setIsThankYouMessageVisible(true);
+	};
+
+	const onTagClick = (e: any) => {
+		setSelectedTag(e.target.value);
+		setIsTagDisabled(true);
+		setTimeout(() => {
+			setAreTagsVisible(false);
+			setIsThankYouMessageVisible(true);
+		}, 1000);
+	};
+
+	return (
+		<div className="flex justify-between w-full pt-4 flex-col gap-4">
+			<div
+				tabIndex={0}
+				className={cn(
+					`rounded-lg text-md w-full text-slate-500 flex justify-between px-4 `,
+				)}
+			>
+				<div className="flex flex-row flex-wrap justify-start">
+					{texts.answerFeedback}
+					<div className="flex sm:mx-2 my-2 sm:my-0">
+						<button
+							onClick={onThumbsUpClick}
+							className={`px-1 text-slate-500 hover:text-parla-blue disabled:hover:text-slate-500`}
+							disabled={isThumbsUpClicked || isThumbsDownClicked}
+						>
+							{isThumbsUpClicked ? <ThumbsUpSolidIcon /> : <ThumbsUpIcon />}
+						</button>
+						<button
+							onClick={onThumbsDownClick}
+							className={`px-1 text-slate-500 hover:text-parla-blue ${isThumbsDownClicked ? "disabled:text-parla-blue" : "disabled:text-slate-500"}  `}
+							disabled={isThumbsDownClicked || isThumbsUpClicked}
+						>
+							{isThumbsDownClicked ? (
+								<ThumbsDownSolidIcon />
+							) : (
+								<ThumbsDownIcon />
+							)}
+						</button>
+					</div>
+				</div>
+				<div className="self-start pt-1 sm:pt-0">
+					<CopyToClipboardButton generatedAnswer={generatedAnswer} />
+				</div>
+			</div>
+
+			<Transition
+				show={areTagsVisible}
+				enter="transition-opacity duration-500"
+				enterFrom="opacity-0"
+				enterTo="opacity-100"
+				leave="transition-opacity duration-100"
+				leaveFrom="opacity-100"
+				leaveTo="opacity-0"
+				className={`flex flex-col self-center rounded-lg shadow-md p-4 bg-white w-full`}
+			>
+				<span className="flex flex-row justify-between">
+					{texts.answerTellUsMore}{" "}
+					<button onClick={() => setAreTagsVisible(false)}>
+						<XIcon className="text-slate-500 hover:text-slate-700 w-5" />
+					</button>
+				</span>
+				<div className="flex flex-wrap gap-x-6 gap-y-4 py-4">
+					{texts?.answerTags.map((tag, idx) => (
+						<label
+							key={idx}
+							className={`rounded-lg border border-slate-300 p-2 
+                        ${selectedTag === tag ? "bg-parla-blue hover:bg-parla-blue text-white" : "hover:bg-slate-50"}`}
+						>
+							<input
+								type="radio"
+								name="tag"
+								onChange={onTagClick}
+								key={idx}
+								value={tag}
+								className="hidden peer"
+								disabled={isTagDisabled}
+							/>
+							{tag}
+						</label>
+					))}
+				</div>
+			</Transition>
+
+			<Transition
+				show={isThankYouMessageVisible}
+				enter="transition-opacity duration-500"
+				enterFrom="opacity-0"
+				enterTo="opacity-100"
+				className={`flex self-center rounded-lg shadow-md p-4 mx-4 bg-white w-fit`}
+			>
+				{texts.answerThankYou}
+			</Transition>
+		</div>
+	);
 }
